OPEN-92 do nothing for empty collections in deleteByIds()

+++ valqueries-automapper/src/main/java/com/valqueries/automapper/ValqueriesAccessDataLayerImpl.java
@@ -108,6 +108,9 @@ public class ValqueriesAccessDataLayerImpl<T, K> implements ValqueriesAccessData
 
 	@Override
 	public CrudUpdateResult deleteByIds(Collection<K> ids) {
+		if (ids.isEmpty()) {
+			return () -> 0;
+		}
 		String inIdsSql = IntStream.range(0, ids.size()).mapToObj(id -> ":" + getKeyName(id)).collect(Collectors.joining(", "));
 		return getUpdateResult(database.obtainInTransaction(tx ->
 			tx.update("DELETE from " + getTableName() +
